Update the codelabs and tests to Angular 1.1. The codelabs were still using the old Controller construct, and they published a Model as a\ separate class. So my updates were more than just changing a few method calls.
In fact, I think we can delete the Controllers code lab and merge it with the
Model lab (and rename Model to something more appropriate).

Also, tests for s4_component and s7_service don't work.

In s4_component:
test 'updates the DOM when the name is changed'
Isn't working. It always reports SUCCESS even when I break it on purpose.
For example, instead of the original:
expect(el.innerHtml, NAME);
I change it to
expect(el.innerHtml, "darren");
Even though NAME is "Misko", the report output says:
Chrome 42.0.2311 (Linux): Executed 3 of 3 SUCCESS (4.364 secs / 0.999 secs)

In s7_service:
 pirate.generateName();
 microLeap();

fails to return a name.